Fixed generation of enum value descriptions
Due to a bug only enum values that are longer than 100 characters
which would in the end generate more than 1 line of description
would be taken into consideration.
This filter, initialy, was done to not produce empty descriptions
but because software development is what it is, it was filtering
everything but the long descriptions.
Fixed.